Initialization phase correctly functioning
+ The ID are correctly added to the Server Administrator.
+ Fixed the error where the added IDs where just replicating, since
  it was using a static field. Now the Taxi class will save their ID in
  a dynamic field inside the main function. While the administrator
  server will save the IDs inside an ArrayList of TaxiInfos (which is a
  class for buffering the data of the Taxi processes).

Issues:
- Now the taxi data structures (TaxiInfo) saved on the administratore server
  won't have updated their taxi list (neighbour taxis, the other taxis
  in the smart city). There will be the need for an updating REST
  service.

+++ DPS_SETA_2022/src/main/java/Servers/ServerThread.java
@@ -30,7 +30,7 @@ public class ServerThread extends Thread {
             taxiId = inFromClient.readLine();
             if (!taxiId.isEmpty()) {
                 int number = Integer.parseInt(taxiId);
-                if (number > 0 && number < Long.MAX_VALUE) {
+                if (number > 0 && number < Integer.MAX_VALUE) {
                     System.out.println(taxiId);
                     Random rand = new Random();
                     int district = rand.nextInt(numberOfDistricts - lowerBound) + lowerBound;
